+++ CHANGELOG.md
@@ -7,6 +7,8 @@ that were not yet released.
 
 _Unreleased_
 
+- Default build-system and license can be specified in global config.  #244
+
 - Fixed an issue where the `init` command would not let you create
   `flit` based projects.  #254